+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/bloomery/ModuleTechBloomeryConfig.java
@@ -4,14 +4,24 @@ import com.codetaylor.mc.athenaeum.parser.recipe.item.MalformedRecipeItemExcepti
 import com.codetaylor.mc.athenaeum.parser.recipe.item.ParseResult;
 import com.codetaylor.mc.athenaeum.parser.recipe.item.RecipeItemParser;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.crafting.Ingredient;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.Tuple;
 import net.minecraftforge.common.config.Config;
 
+import java.util.ArrayList;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
 
 @Config(modid = ModuleTechBloomery.MOD_ID, name = ModuleTechBloomery.MOD_ID + "/" + "module.tech.Bloomery")
 public class ModuleTechBloomeryConfig {
+
+  @Config.Ignore
+  public static final List<Tuple<Ingredient, Double>> BLOOMERY_FUEL_MODIFIERS = new ArrayList<>(1);
+
+  @Config.Ignore
+  public static final List<Tuple<Ingredient, Double>> WITHER_FORGE_FUEL_MODIFIERS = new ArrayList<>(1);
 
   // ---------------------------------------------------------------------------
   // - Slag
@@ -140,6 +150,18 @@ public class ModuleTechBloomeryConfig {
     }};
 
     public double getSpecialFuelBurnTimeModifier(ItemStack stack) {
+
+      // Check the tuple list first
+
+      for (Tuple<Ingredient, Double> modifier : BLOOMERY_FUEL_MODIFIERS) {
+        Ingredient ingredient = modifier.getFirst();
+
+        if (ingredient.apply(stack)) {
+          return modifier.getSecond();
+        }
+      }
+
+      // Next, check config values
 
       ResourceLocation registryName = stack.getItem().getRegistryName();
 
@@ -281,6 +303,18 @@ public class ModuleTechBloomeryConfig {
     }};
 
     public double getSpecialFuelBurnTimeModifier(ItemStack stack) {
+
+      // Check the tuple list first
+
+      for (Tuple<Ingredient, Double> modifier : WITHER_FORGE_FUEL_MODIFIERS) {
+        Ingredient ingredient = modifier.getFirst();
+
+        if (ingredient.apply(stack)) {
+          return modifier.getSecond();
+        }
+      }
+
+      // Next, check config values
 
       ResourceLocation registryName = stack.getItem().getRegistryName();