+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/bloomery/plugin/crafttweaker/ZenBloomery.java
@@ -8,6 +8,7 @@ import com.codetaylor.mc.athenaeum.tools.ZenDocMethod;
 import com.codetaylor.mc.pyrotech.library.crafttweaker.RemoveAllRecipesAction;
 import com.codetaylor.mc.pyrotech.modules.tech.basic.recipe.AnvilRecipe;
 import com.codetaylor.mc.pyrotech.modules.tech.bloomery.ModuleTechBloomery;
+import com.codetaylor.mc.pyrotech.modules.tech.bloomery.ModuleTechBloomeryConfig;
 import com.codetaylor.mc.pyrotech.modules.tech.bloomery.recipe.*;
 import crafttweaker.IAction;
 import crafttweaker.api.item.IIngredient;
@@ -16,6 +17,7 @@ import crafttweaker.api.minecraft.CraftTweakerMC;
 import crafttweaker.mc1120.CraftTweaker;
 import net.minecraft.item.crafting.Ingredient;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.Tuple;
 import stanhebben.zenscript.annotations.ZenClass;
 import stanhebben.zenscript.annotations.ZenMethod;
 
@@ -133,6 +135,42 @@ public class ZenBloomery {
             CTInputHelper.toIngredient(input)
         ),
         EnumRecipeType.WitherForge
+    );
+  }
+
+  @ZenDocMethod(
+      order = 5,
+      args = {
+          @ZenDocArg(arg = "fuel"),
+          @ZenDocArg(arg = "modifier")
+      }
+  )
+  @ZenMethod
+  public static void addBloomeryFuelModifier(IIngredient fuel, double modifier) {
+
+    ModuleTechBloomeryConfig.BLOOMERY_FUEL_MODIFIERS.add(
+        new Tuple<>(
+            CTInputHelper.toIngredient(fuel),
+            modifier
+        )
+    );
+  }
+
+  @ZenDocMethod(
+      order = 6,
+      args = {
+          @ZenDocArg(arg = "fuel"),
+          @ZenDocArg(arg = "modifier")
+      }
+  )
+  @ZenMethod
+  public static void addWitherForgeFuelModifier(IIngredient fuel, double modifier) {
+
+    ModuleTechBloomeryConfig.WITHER_FORGE_FUEL_MODIFIERS.add(
+        new Tuple<>(
+            CTInputHelper.toIngredient(fuel),
+            modifier
+        )
     );
   }